scripts/rai/requirements.txt
azure-identity==1.21.0
python-dotenv==1.1.0
azure-mgmt-resource==23.3.0